Implement import deck to frontend

See merge request swtp-suess/ws-23/card-trainer!65

+++ app/frontend/angular-frontend/src/app/deck-list-view/deck-list-view.component.ts
@@ -8,6 +8,7 @@ import {FormsModule} from "@angular/forms";
 import {ToastService} from "../services/toast-service/toast.service";
 import {ToasterComponent} from "../toaster/toaster.component";
 import {AuthService} from "../services/auth-service/auth-service";
+import {NgbModal, NgbModalRef} from "@ng-bootstrap/ng-bootstrap";
 
 @Component({
   selector: 'app-deck-list-view',
@@ -21,13 +22,23 @@ export class DeckListViewComponent implements OnInit {
   deckList: DeckDTO[] = [];
   showOptions: boolean = false;
   showDelete: boolean = false;
+  private modalRef: NgbModalRef | undefined;
+  buttonIsPressed: boolean = false;
+  importFile: string = "";
+  private selectedFile: File | null = null;
 
   toggleOptions() {
     this.showOptions = !this.showOptions;
     document.getElementById("")?.classList.toggle("show")
   }
 
-  constructor(private cardService: CardService, private router: Router, private toast: ToastService, private userService: AuthService) {
+  constructor(
+    private cardService: CardService,
+    private router: Router,
+    private toast: ToastService,
+    private userService: AuthService,
+    private modalService: NgbModal
+  ) {
   }
 
   ngOnInit() {
@@ -53,11 +64,18 @@ export class DeckListViewComponent implements OnInit {
   }
 
   addItem() {
+    // Added this to avoid having multible 'Enter new Deck name here'-Field
+    if (this.deckList.at(-1)?.deckId == -1) {
+      this.deckList.pop()
+      return
+    }
+
     const deck1: DeckDTO = {
       deckName: "NewItem",
       deckId: -1,
       cardsToLearn: 0
     }
+
     this.deckList.push(deck1)
   }
 
@@ -77,6 +95,8 @@ export class DeckListViewComponent implements OnInit {
 
   deleteItem($event: Event, id: number) {
     $event.stopPropagation()
+    if (!confirm("Are you sure you want to delete the deck?"))
+      return
     this.cardService.deleteDeck(id).subscribe({
       complete: () => {
         this.deckList = this.deckList.filter(value => value.deckId != id);
@@ -86,5 +106,34 @@ export class DeckListViewComponent implements OnInit {
 
   toggleDeleteItem() {
     this.showDelete = !this.showDelete
+  }
+
+  onSubmitImportDeck(event: any) {
+
+    if (!this.selectedFile) {
+      return;
+    }
+
+    const formData = new FormData();
+    formData.append('file', this.selectedFile, this.selectedFile.name);
+    this.cardService.importDeckUpload(formData).subscribe({
+      next: value => {
+        this.toast.showSuccessToast("Import Deck", "Your deck has been imported successfully");
+        this.updateDecks();
+        this.modalRef?.close()
+      },
+      error: err => {
+        this.toast.showErrorToast("Import deck", "Your deck could not be imported. Please make sure you are using a valid deck.zip 😊")
+      }
+    });
+  }
+
+  showImportModal(content: any) {
+    this.modalRef = this.modalService.open(content);
+  }
+
+  onFileSelected(event: any) {
+    this.selectedFile = event.target.files[0];
+    console.log(this.selectedFile)
   }
 }
